progress

+++ chain/common_tx.go
@@ -48,8 +48,8 @@ func updateSpace(s string, t *TransactionContext, timeRemaining uint64, i *Space
 	return PutSpaceInfo(t.Database, []byte(s), i, lastExpiry)
 }
 
-func valueUnits(g *Genesis, size int) uint64 {
-	return uint64(size)/g.ValueUnitSize + 1
+func valueUnits(g *Genesis, size uint64) uint64 {
+	return size/g.ValueUnitSize + 1
 }
 
 func valueHash(v []byte) string {
